+++ pkg/autodetect/manifestd/collectors/common.go
