+++ pkg/autodetect/manifestd/collectors/denom_info.go
