+++ pkg/autodetect/manifestd/collectors/denom_info_umfx.go
