+++ pkg/autodetect/manifestd/collectors/registry.go
