+++ pkg/autodetect/manifestd/collectors/token_count.go
