+++ pkg/autodetect/registry.go
@@ -1,4 +1,4 @@
-package monitors
+package autodetect
 
 import (
 	"context"
