+++ pkg/autodetect/utils.go
